+++ src/include/smash/particledata.h
@@ -1,5 +1,5 @@
 /*
- *    Copyright (c) 2012-2020,2022-2023
+ *    Copyright (c) 2012-2020,2022-2024
  *      SMASH Team
  *
  *    GNU General Public License (GPLv3 or later)
@@ -575,7 +575,7 @@ std::ostream &operator<<(std::ostream &out,
  * \param[in] log_area Logging area for the warning
  * \param[inout] mass_warning Whether to warn about mass discrepancies
  * \param[inout] on_shell_warning Whether to warn about off-shell particles
- * \return
+ * \return Valid SMASH particle matching all given input quantities
  *
  * \note The boolean flags are passed by reference, since we want to allow
  * client code to warn the user only once per flag. Hence, this function is