+++ CHANGELOG.md
@@ -20,15 +20,18 @@ Also possible, but for this project less relevant, is `Deprecated` for soon-to-b
 
 ## Unreleased
 
-## SMASH-3.1rc
-Date:
+
+## SMASH-3.1
+Date: 2024-02-26
 
 ### Added
 * Added momentum dependence of the Skyrme potential using the cold nuclear matter approximation
 * Option `Decay_Initial_Particles` which blocks the decay of resonances created at initialization
 * The `List` and `ListBox` modi can now be used specifying a single particles input file using the `Filename` key
 * Command-line option to run smash with `-n` to avoid caching integrals in the ***tabulations*** folder on the disk
-* ⚠️  The `Total_Cross_Section_Strategy` key determines how the total cross sections for collision finding are evaluated, whether from summing the partial contributions or using a parametrization
+* ⚠️  The `Total_Cross_Section_Strategy` key determines how the total cross sections for collision finding are evaluated, whether from summing the partial contributions or using a parametrization; previously the default behavior was the former strategy (now accessible with the value `BottomUp`), but it has changed to the latter (accessible with `TopDownMeasured`)
+* ⚠️  Option `Pseudoresonance` which adds *ad hoc* inelastic 2→1 processes close to the energy regime around the transition to strings, where few resonances have been measured; by default, this is done for interactions where at least one of the incoming particles is unstable, and the heaviest possible resonance producible from the incoming pair is chosen
+* New `Use_Potentials_Outside_Lattice` option in the config file to specify where potentials are applied
 
 ### Fixed
 * Fix bug in the time step calculation of SMASH output in events successive to the first one when the `End_Time` is not a multiple of the `Output_Interval`
@@ -38,6 +41,9 @@ Date:
 * Fix value of particles mass in SMASH IC ASCII output (the pole mass was printed instead of the real mass)
 * SMASH does not crash if `rho_eckart` is not specified in the `Thermodynamics` output
 * SMASH does not crash if the same output format is repeated
+* Fix bug when `Power_Particle_Formation` is explicitly specified
+* SMASH can now be built with GNU compiler `v13` that requires explicit inclusion of `cstdint` header
+* Fix bug when using SMASH as a library and particles should be removed
 
 ### Changed
 * Use updated gold radius of 6.55 fm instead of 6.38 fm and a diffusiveness of 0.523 fm instead of 0.535 fm